list_product_fix_delete_admin_25_04

+++ db/migrate/20200424025733_create_products.rb
@@ -3,8 +3,8 @@ class CreateProducts < ActiveRecord::Migration[6.0]
     create_table :products do |t|
       t.string :pro_name
       t.integer :quantity
-      t.float :price_input
-      t.float :price_output
+      t.decimal :price_input
+      t.decimal :price_output
       t.string :img_1
       t.string :img_2
       t.string :img_3
